Refactor styles for Finance Admin Screen and add common styles for various screens
- Updated import paths for FinanceAdminScreen styles.
- Created common screen styles in `screen.css` for shared layout and design.
- Added modal component styles in `modals.css` for consistent modal design.
- Introduced specific styles for Finance Admin Screen in `FinanceAdminScreen.css`.
- Added specific styles for Cycles, Dashboard, Finance Admins, Guards, and Students screens.
- Enhanced loading states and responsive design across various screens.
